// src/app/models/source.ts
export class SourceDetails extends BaseModel {
    id: string;
    sourcingId?: string;
    dateEntered: Date;
    dateModified: Date;
    createdById: string;
    modifiedById: string;
    assignedSalesRepId: string;
    itemName: string;
    itemNumber: string;
    gcName: string;
    gcItemNumber: string;
    size: string;
    material: string;
    specifications: string;
    decorationType: string;
    testingRequirements: string;
    artCharge: number;
    toolingCharge: number;
    toolingTime: number;
    sampleCost: number;
    sampleTime: number;
    productionLeadTime: number;
    packaging: string;
    masterCartonWidth: string;
    masterCartonHeight: string;
    masterCartonDepth: string;
    masterCartonWeight: string;
    fobChinaPort: string;
    factoryTerms: string;
    factoryRemarks: string;
    thirdPartyAudits: string; 
    quoteValidThrough: Date;
    certificates: string;
    status: string;

    createdByName: string;
    modifiedByName: string;
    assignedSalesRep: string;
    images: string[];

    accountName: string;
    accountId: string;
    contactName: string;
    contactId: string;
    quantity1: number;
    pricePerQty1: number;
    quantity2: number;
    pricePerQty2: number;
    quantity3: number;
    pricePerQty3: number;
    quantity4: number;
    pricePerQty4: number;
    quantity5: number;
    pricePerQty5: number;

    opportunityId: string;
    opportunityNo: string;
    assignedSourcingRep: string;
    assignedSourcingRepId: string;

    setData(data) {
        this.setString(data, 'id');
        this.setString(data, 'sourcingId');
        this.setDate(data, 'dateEntered');
        this.setDate(data, 'dateModified');
        this.setString(data, 'createdById');
        this.setString(data, 'modifiedById');
        this.setString(data, 'itemName');
        this.setString(data, 'itemNumber');
        this.setString(data, 'gcName');
        this.setString(data, 'gcItemNumber', FuseUtils.generateGUID());
        this.setString(data, 'size');
        this.setString(data, 'material');
        this.setString(data, 'specifications');
        this.setString(data, 'decorationType');
        this.setString(data, 'testingRequirements');
        this.setFloat(data, 'artCharge');
        this.setFloat(data, 'toolingCharge');
        this.setInt(data, 'toolingTime');
        this.setInt(data, 'sampleCost');
        this.setInt(data, 'sampleTime');
        this.setInt(data, 'productionLeadTime');
        this.setString(data, 'packaging');
        this.setString(data, 'masterCartonWidth');
        this.setString(data, 'masterCartonHeight');
        this.setString(data, 'masterCartonDepth');
        this.setString(data, 'masterCartonWeight');
        this.setString(data, 'fobChinaPort');
        this.setString(data, 'factoryTerms');
        this.setString(data, 'factoryRemarks');
        this.setString(data, 'thirdPartyAudits');
        this.setDate(data, 'quoteValidThrough');
        this.setString(data, 'certificates');
        this.setString(data, 'status');
        this.setString(data, 'createdByName');
        this.setString(data, 'modifiedByName');
        this.setString(data, 'assignedSalesRep');
        this.setString(data, 'assignedSalesRepId');
        this.setString(data, 'assignedSourcingRep');
        this.setString(data, 'assignedSourcingRepId');
        this.setArray(data, 'images');
        this.setString(data, 'accountName');
        this.setString(data, 'accountId');
        this.setString(data, 'contactName');
        this.setString(data, 'contactId');
        this.setInt(data, 'quantity1');
        this.setFloat(data, 'pricePerQty1');
        this.setInt(data, 'quantity2');
        this.setFloat(data, 'pricePerQty2');
        this.setInt(data, 'quantity3');
        this.setFloat(data, 'pricePerQty3');
        this.setInt(data, 'quantity4');
        this.setFloat(data, 'pricePerQty4');
        this.setInt(data, 'quantity5');
        this.setFloat(data, 'pricePerQty5');
        this.setString(data, 'opportunityId');
        this.setString(data, 'opportunityNo');
    }
}